+++ rstar/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Added
 - Add optional support for the [mint](https://docs.rs/mint/0.5.9/mint/index.html) crate
+- Implemented `IntoIter` for `RTree`, i.e. added a owning iterator
 - Added cached envelope bulk load benchmark
 
 ## Changed